MISRA Rul2 10.4.: Aligning types to the same essential type category

+++ hal/armv7r/zynqmp/zynqmp.c
@@ -19,7 +19,7 @@
 #include "include/arch/armv7r/zynqmp/zynqmp.h"
 
 #include "hal/armv7r/halsyspage.h"
-#include "zynqmp_regs.h"/*TBC - hal/armv7r/zynqmp_regs.h*/
+#include "zynqmp_regs.h" /*TBC - hal/armv7r/zynqmp_regs.h*/
 
 #define IOU_SLCR_BASE_ADDRESS 0xff180000U
 #define APU_BASE_ADDRESS      0xfd5c0000U
@@ -41,7 +41,7 @@ struct {
 static int _zynqmp_setBasicGenerator(volatile u32 *reg, u32 dev, u8 src, u8 div0, u8 div1, u8 active)
 {
 	u32 val = src;
-	if (dev == pctl_devclock_lpd_timestamp) {
+	if (dev == (unsigned int)pctl_devclock_lpd_timestamp) {
 		val &= 0x7U;
 	}
 	else {
@@ -49,7 +49,7 @@ static int _zynqmp_setBasicGenerator(volatile u32 *reg, u32 dev, u8 src, u8 div0
 	}
 
 	val |= (((u32)div0 & 0x3fU) << 8) | (((u32)div1 & 0x3fU) << 16) | ((u32)active << 24);
-	if (dev == pctl_devclock_lpd_cpu_r5) {
+	if (dev == (unsigned int)pctl_devclock_lpd_cpu_r5) {
 		/* According to docs turning this bit off could lead to system hang - ensure it is on */
 		val |= (0x01UL << 24);
 	}
@@ -62,12 +62,12 @@ static int _zynqmp_setBasicGenerator(volatile u32 *reg, u32 dev, u8 src, u8 div0
 
 int _zynqmp_setDevClock(u32 dev, u8 src, u8 div0, u8 div1, u8 active)
 {
-	if ((dev >= pctl_devclock_lpd_usb3_dual) && (dev <= pctl_devclock_lpd_timestamp)) {
-		unsigned regOffset = (dev - pctl_devclock_lpd_usb3_dual) + crl_apb_usb3_dual_ref_ctrl;
+	if ((dev >= (unsigned int)pctl_devclock_lpd_usb3_dual) && (dev <= (unsigned int)pctl_devclock_lpd_timestamp)) {
+		unsigned regOffset = (dev - (unsigned int)pctl_devclock_lpd_usb3_dual) + (unsigned int)crl_apb_usb3_dual_ref_ctrl;
 		return _zynqmp_setBasicGenerator(zynq_common.crl_apb + regOffset, dev, src, div0, div1, active);
 	}
 	else if ((dev >= pctl_devclock_fpd_acpu) && (dev <= pctl_devclock_fpd_dbg_tstmp)) {
-		unsigned regOffset = (dev - pctl_devclock_fpd_acpu) + crf_apb_acpu_ctrl;
+		unsigned regOffset = (dev - (unsigned int)pctl_devclock_fpd_acpu) + crf_apb_acpu_ctrl;
 		return _zynqmp_setBasicGenerator(zynq_common.crf_apb + regOffset, dev, src, div0, div1, active);
 	}
 
@@ -88,12 +88,12 @@ static int _zynqmp_getBasicGenerator(volatile u32 *reg, u8 *src, u8 *div0, u8 *d
 
 int _zynqmp_getDevClock(u32 dev, u8 *src, u8 *div0, u8 *div1, u8 *active)
 {
-	if ((dev >= pctl_devclock_lpd_usb3_dual) && (dev <= pctl_devclock_lpd_timestamp)) {
-		unsigned regOffset = (dev - pctl_devclock_lpd_usb3_dual) + crl_apb_usb3_dual_ref_ctrl;
+	if ((dev >= (unsigned int)pctl_devclock_lpd_usb3_dual) && (dev <= (unsigned int)pctl_devclock_lpd_timestamp)) {
+		unsigned regOffset = (dev - (unsigned int)pctl_devclock_lpd_usb3_dual) + (unsigned int)crl_apb_usb3_dual_ref_ctrl;
 		return _zynqmp_getBasicGenerator(zynq_common.crl_apb + regOffset, src, div0, div1, active);
 	}
 	else if ((dev >= pctl_devclock_fpd_acpu) && (dev <= pctl_devclock_fpd_dbg_tstmp)) {
-		unsigned regOffset = (dev - pctl_devclock_fpd_acpu) + crf_apb_acpu_ctrl;
+		unsigned regOffset = (dev - (unsigned int)pctl_devclock_fpd_acpu) + (unsigned int)crf_apb_acpu_ctrl;
 		return _zynqmp_getBasicGenerator(zynq_common.crf_apb + regOffset, src, div0, div1, active);
 	}
 
@@ -110,7 +110,7 @@ static void _zynqmp_setMIOMuxing(unsigned pin, u8 l0, u8 l1, u8 l2, u8 l3)
 
 static void _zynqmp_setMIOTristate(unsigned pin, u8 config)
 {
-	u32 reg = pin / 32U + iou_slcr_mio_mst_tri0;
+	u32 reg = pin / 32U + (unsigned int)iou_slcr_mio_mst_tri0;
 	u32 bit = pin % 32U;
 	u32 mask = 0x01UL << bit;
 
@@ -124,7 +124,7 @@ static void _zynqmp_setMIOTristate(unsigned pin, u8 config)
 
 static void _zynqmp_setMIOControl(unsigned pin, u8 config)
 {
-	u32 reg = (pin / 26U) * (iou_slcr_bank1_ctrl0 - iou_slcr_bank0_ctrl0) + iou_slcr_bank0_ctrl0;
+	u32 reg = (pin / 26U) * (unsigned int)(iou_slcr_bank1_ctrl0 - iou_slcr_bank0_ctrl0) + (unsigned int)iou_slcr_bank0_ctrl0;
 	u32 bit = pin % 26U;
 	u32 mask = 0x01UL << bit;
 	unsigned int i;
@@ -171,7 +171,7 @@ static void _zynqmp_getMIOMuxing(unsigned pin, u8 *l0, u8 *l1, u8 *l2, u8 *l3)
 
 static void _zynqmp_getMIOTristate(unsigned pin, u8 *config)
 {
-	u32 reg = pin / 32U + iou_slcr_mio_mst_tri0;
+	u32 reg = pin / 32U + (unsigned int)iou_slcr_mio_mst_tri0;
 	u32 bit = pin % 32U;
 	if ((*(zynq_common.iou_slcr + reg) & (0x01U << bit)) != 0U) {
 		*config |= PCTL_MIO_TRI_ENABLE;
@@ -180,7 +180,7 @@ static void _zynqmp_getMIOTristate(unsigned pin, u8 *config)
 
 static void _zynqmp_getMIOControl(unsigned pin, u8 *config)
 {
-	u32 reg = (pin / 26U) * (iou_slcr_bank1_ctrl0 - iou_slcr_bank0_ctrl0) + iou_slcr_bank0_ctrl0;
+	u32 reg = (pin / 26U) * (unsigned int)(iou_slcr_bank1_ctrl0 - iou_slcr_bank0_ctrl0) + (unsigned int)iou_slcr_bank0_ctrl0;
 	u32 bit = pin % 26U;
 	u32 mask = 0x01UL << bit;
 	unsigned int i;
@@ -216,82 +216,82 @@ static int _zynqmp_getMIO(unsigned pin, u8 *l0, u8 *l1, u8 *l2, u8 *l3, u8 *conf
 static int _zynqmp_parseReset(int dev, volatile u32 **reg, u32 *bit)
 {
 	static const u32 lookup[] = {
-		[pctl_devreset_lpd_gem0] = crl_apb_rst_lpd_iou0 | (0UL << 12),
-		[pctl_devreset_lpd_gem1] = crl_apb_rst_lpd_iou0 | (1UL << 12),
-		[pctl_devreset_lpd_gem2] = crl_apb_rst_lpd_iou0 | (2UL << 12),
-		[pctl_devreset_lpd_gem3] = crl_apb_rst_lpd_iou0 | (3UL << 12),
-		[pctl_devreset_lpd_qspi] = crl_apb_rst_lpd_iou2 | (0UL << 12),
-		[pctl_devreset_lpd_uart0] = crl_apb_rst_lpd_iou2 | (1UL << 12),
-		[pctl_devreset_lpd_uart1] = crl_apb_rst_lpd_iou2 | (2UL << 12),
-		[pctl_devreset_lpd_spi0] = crl_apb_rst_lpd_iou2 | (3UL << 12),
-		[pctl_devreset_lpd_spi1] = crl_apb_rst_lpd_iou2 | (4UL << 12),
-		[pctl_devreset_lpd_sdio0] = crl_apb_rst_lpd_iou2 | (5UL << 12),
-		[pctl_devreset_lpd_sdio1] = crl_apb_rst_lpd_iou2 | (6UL << 12),
-		[pctl_devreset_lpd_can0] = crl_apb_rst_lpd_iou2 | (7UL << 12),
-		[pctl_devreset_lpd_can1] = crl_apb_rst_lpd_iou2 | (8UL << 12),
-		[pctl_devreset_lpd_i2c0] = crl_apb_rst_lpd_iou2 | (9UL << 12),
-		[pctl_devreset_lpd_i2c1] = crl_apb_rst_lpd_iou2 | (10UL << 12),
-		[pctl_devreset_lpd_ttc0] = crl_apb_rst_lpd_iou2 | (11UL << 12),
-		[pctl_devreset_lpd_ttc1] = crl_apb_rst_lpd_iou2 | (12UL << 12),
-		[pctl_devreset_lpd_ttc2] = crl_apb_rst_lpd_iou2 | (13UL << 12),
-		[pctl_devreset_lpd_ttc3] = crl_apb_rst_lpd_iou2 | (14UL << 12),
-		[pctl_devreset_lpd_swdt] = crl_apb_rst_lpd_iou2 | (15UL << 12),
-		[pctl_devreset_lpd_nand] = crl_apb_rst_lpd_iou2 | (16UL << 12),
-		[pctl_devreset_lpd_lpd_dma] = crl_apb_rst_lpd_iou2 | (17UL << 12),
-		[pctl_devreset_lpd_gpio] = crl_apb_rst_lpd_iou2 | (18UL << 12),
-		[pctl_devreset_lpd_iou_cc] = crl_apb_rst_lpd_iou2 | (19UL << 12),
-		[pctl_devreset_lpd_timestamp] = crl_apb_rst_lpd_iou2 | (20UL << 12),
-		[pctl_devreset_lpd_rpu_r50] = crl_apb_rst_lpd_top | (0UL << 12),
-		[pctl_devreset_lpd_rpu_r51] = crl_apb_rst_lpd_top | (1UL << 12),
-		[pctl_devreset_lpd_rpu_amba] = crl_apb_rst_lpd_top | (2UL << 12),
-		[pctl_devreset_lpd_ocm] = crl_apb_rst_lpd_top | (3UL << 12),
-		[pctl_devreset_lpd_rpu_pge] = crl_apb_rst_lpd_top | (4UL << 12),
-		[pctl_devreset_lpd_usb0_corereset] = crl_apb_rst_lpd_top | (6UL << 12),
-		[pctl_devreset_lpd_usb1_corereset] = crl_apb_rst_lpd_top | (7UL << 12),
-		[pctl_devreset_lpd_usb0_hiberreset] = crl_apb_rst_lpd_top | (8UL << 12),
-		[pctl_devreset_lpd_usb1_hiberreset] = crl_apb_rst_lpd_top | (9UL << 12),
-		[pctl_devreset_lpd_usb0_apb] = crl_apb_rst_lpd_top | (10UL << 12),
-		[pctl_devreset_lpd_usb1_apb] = crl_apb_rst_lpd_top | (11UL << 12),
-		[pctl_devreset_lpd_ipi] = crl_apb_rst_lpd_top | (14UL << 12),
-		[pctl_devreset_lpd_apm] = crl_apb_rst_lpd_top | (15UL << 12),
-		[pctl_devreset_lpd_rtc] = crl_apb_rst_lpd_top | (16UL << 12),
-		[pctl_devreset_lpd_sysmon] = crl_apb_rst_lpd_top | (17UL << 12),
-		[pctl_devreset_lpd_s_axi_lpd] = crl_apb_rst_lpd_top | (19UL << 12),
-		[pctl_devreset_lpd_lpd_swdt] = crl_apb_rst_lpd_top | (20UL << 12),
-		[pctl_devreset_lpd_fpd] = crl_apb_rst_lpd_top | (23UL << 12),
-		[pctl_devreset_lpd_dbg_fpd] = crl_apb_rst_lpd_dbg | (0UL << 12),
-		[pctl_devreset_lpd_dbg_lpd] = crl_apb_rst_lpd_dbg | (1UL << 12),
-		[pctl_devreset_lpd_rpu_dbg0] = crl_apb_rst_lpd_dbg | (4UL << 12),
-		[pctl_devreset_lpd_rpu_dbg1] = crl_apb_rst_lpd_dbg | (5UL << 12),
-		[pctl_devreset_lpd_dbg_ack] = crl_apb_rst_lpd_dbg | (15UL << 12),
-		[pctl_devreset_fpd_sata] = crf_apb_rst_fpd_top | (1UL << 12),
-		[pctl_devreset_fpd_gt] = crf_apb_rst_fpd_top | (2UL << 12),
-		[pctl_devreset_fpd_gpu] = crf_apb_rst_fpd_top | (3UL << 12),
-		[pctl_devreset_fpd_gpu_pp0] = crf_apb_rst_fpd_top | (4UL << 12),
-		[pctl_devreset_fpd_gpu_pp1] = crf_apb_rst_fpd_top | (5UL << 12),
-		[pctl_devreset_fpd_fpd_dma] = crf_apb_rst_fpd_top | (6UL << 12),
-		[pctl_devreset_fpd_s_axi_hpc_0_fpd] = crf_apb_rst_fpd_top | (7UL << 12),
-		[pctl_devreset_fpd_s_axi_hpc_1_fpd] = crf_apb_rst_fpd_top | (8UL << 12),
-		[pctl_devreset_fpd_s_axi_hp_0_fpd] = crf_apb_rst_fpd_top | (9UL << 12),
-		[pctl_devreset_fpd_s_axi_hp_1_fpd] = crf_apb_rst_fpd_top | (10UL << 12),
-		[pctl_devreset_fpd_s_axi_hpc_2_fpd] = crf_apb_rst_fpd_top | (11UL << 12),
-		[pctl_devreset_fpd_s_axi_hpc_3_fpd] = crf_apb_rst_fpd_top | (12UL << 12),
-		[pctl_devreset_fpd_swdt] = crf_apb_rst_fpd_top | (15UL << 12),
-		[pctl_devreset_fpd_dp] = crf_apb_rst_fpd_top | (16UL << 12),
-		[pctl_devreset_fpd_pcie_ctrl] = crf_apb_rst_fpd_top | (17UL << 12),
-		[pctl_devreset_fpd_pcie_bridge] = crf_apb_rst_fpd_top | (18UL << 12),
-		[pctl_devreset_fpd_pcie_cfg] = crf_apb_rst_fpd_top | (19UL << 12),
-		[pctl_devreset_fpd_acpu0] = crf_apb_rst_fpd_apu | (0UL << 12),
-		[pctl_devreset_fpd_acpu1] = crf_apb_rst_fpd_apu | (1UL << 12),
-		[pctl_devreset_fpd_acpu2] = crf_apb_rst_fpd_apu | (2UL << 12),
-		[pctl_devreset_fpd_acpu3] = crf_apb_rst_fpd_apu | (3UL << 12),
-		[pctl_devreset_fpd_apu_l2] = crf_apb_rst_fpd_apu | (8UL << 12),
-		[pctl_devreset_fpd_acpu0_pwron] = crf_apb_rst_fpd_apu | (10UL << 12),
-		[pctl_devreset_fpd_acpu1_pwron] = crf_apb_rst_fpd_apu | (11UL << 12),
-		[pctl_devreset_fpd_acpu2_pwron] = crf_apb_rst_fpd_apu | (12UL << 12),
-		[pctl_devreset_fpd_acpu3_pwron] = crf_apb_rst_fpd_apu | (13UL << 12),
-		[pctl_devreset_fpd_ddr_apm] = crf_apb_rst_ddr_ss | (2UL << 12),
-		[pctl_devreset_fpd_ddr_reserved] = crf_apb_rst_ddr_ss | (3UL << 12),
+		[pctl_devreset_lpd_gem0] = (unsigned)crl_apb_rst_lpd_iou0 | (0UL << 12),
+		[pctl_devreset_lpd_gem1] = (unsigned)crl_apb_rst_lpd_iou0 | (1UL << 12),
+		[pctl_devreset_lpd_gem2] = (unsigned)crl_apb_rst_lpd_iou0 | (2UL << 12),
+		[pctl_devreset_lpd_gem3] = (unsigned)crl_apb_rst_lpd_iou0 | (3UL << 12),
+		[pctl_devreset_lpd_qspi] = (unsigned)crl_apb_rst_lpd_iou2 | (0UL << 12),
+		[pctl_devreset_lpd_uart0] = (unsigned)crl_apb_rst_lpd_iou2 | (1UL << 12),
+		[pctl_devreset_lpd_uart1] = (unsigned)crl_apb_rst_lpd_iou2 | (2UL << 12),
+		[pctl_devreset_lpd_spi0] = (unsigned)crl_apb_rst_lpd_iou2 | (3UL << 12),
+		[pctl_devreset_lpd_spi1] = (unsigned)crl_apb_rst_lpd_iou2 | (4UL << 12),
+		[pctl_devreset_lpd_sdio0] = (unsigned)crl_apb_rst_lpd_iou2 | (5UL << 12),
+		[pctl_devreset_lpd_sdio1] = (unsigned)crl_apb_rst_lpd_iou2 | (6UL << 12),
+		[pctl_devreset_lpd_can0] = (unsigned)crl_apb_rst_lpd_iou2 | (7UL << 12),
+		[pctl_devreset_lpd_can1] = (unsigned)crl_apb_rst_lpd_iou2 | (8UL << 12),
+		[pctl_devreset_lpd_i2c0] = (unsigned)crl_apb_rst_lpd_iou2 | (9UL << 12),
+		[pctl_devreset_lpd_i2c1] = (unsigned)crl_apb_rst_lpd_iou2 | (10UL << 12),
+		[pctl_devreset_lpd_ttc0] = (unsigned)crl_apb_rst_lpd_iou2 | (11UL << 12),
+		[pctl_devreset_lpd_ttc1] = (unsigned)crl_apb_rst_lpd_iou2 | (12UL << 12),
+		[pctl_devreset_lpd_ttc2] = (unsigned)crl_apb_rst_lpd_iou2 | (13UL << 12),
+		[pctl_devreset_lpd_ttc3] = (unsigned)crl_apb_rst_lpd_iou2 | (14UL << 12),
+		[pctl_devreset_lpd_swdt] = (unsigned)crl_apb_rst_lpd_iou2 | (15UL << 12),
+		[pctl_devreset_lpd_nand] = (unsigned)crl_apb_rst_lpd_iou2 | (16UL << 12),
+		[pctl_devreset_lpd_lpd_dma] = (unsigned)crl_apb_rst_lpd_iou2 | (17UL << 12),
+		[pctl_devreset_lpd_gpio] = (unsigned)crl_apb_rst_lpd_iou2 | (18UL << 12),
+		[pctl_devreset_lpd_iou_cc] = (unsigned)crl_apb_rst_lpd_iou2 | (19UL << 12),
+		[pctl_devreset_lpd_timestamp] = (unsigned)crl_apb_rst_lpd_iou2 | (20UL << 12),
+		[pctl_devreset_lpd_rpu_r50] = (unsigned)crl_apb_rst_lpd_top | (0UL << 12),
+		[pctl_devreset_lpd_rpu_r51] = (unsigned)crl_apb_rst_lpd_top | (1UL << 12),
+		[pctl_devreset_lpd_rpu_amba] = (unsigned)crl_apb_rst_lpd_top | (2UL << 12),
+		[pctl_devreset_lpd_ocm] = (unsigned)crl_apb_rst_lpd_top | (3UL << 12),
+		[pctl_devreset_lpd_rpu_pge] = (unsigned)crl_apb_rst_lpd_top | (4UL << 12),
+		[pctl_devreset_lpd_usb0_corereset] = (unsigned)crl_apb_rst_lpd_top | (6UL << 12),
+		[pctl_devreset_lpd_usb1_corereset] = (unsigned)crl_apb_rst_lpd_top | (7UL << 12),
+		[pctl_devreset_lpd_usb0_hiberreset] = (unsigned)crl_apb_rst_lpd_top | (8UL << 12),
+		[pctl_devreset_lpd_usb1_hiberreset] = (unsigned)crl_apb_rst_lpd_top | (9UL << 12),
+		[pctl_devreset_lpd_usb0_apb] = (unsigned)crl_apb_rst_lpd_top | (10UL << 12),
+		[pctl_devreset_lpd_usb1_apb] = (unsigned)crl_apb_rst_lpd_top | (11UL << 12),
+		[pctl_devreset_lpd_ipi] = (unsigned)crl_apb_rst_lpd_top | (14UL << 12),
+		[pctl_devreset_lpd_apm] = (unsigned)crl_apb_rst_lpd_top | (15UL << 12),
+		[pctl_devreset_lpd_rtc] = (unsigned)crl_apb_rst_lpd_top | (16UL << 12),
+		[pctl_devreset_lpd_sysmon] = (unsigned)crl_apb_rst_lpd_top | (17UL << 12),
+		[pctl_devreset_lpd_s_axi_lpd] = (unsigned)crl_apb_rst_lpd_top | (19UL << 12),
+		[pctl_devreset_lpd_lpd_swdt] = (unsigned)crl_apb_rst_lpd_top | (20UL << 12),
+		[pctl_devreset_lpd_fpd] = (unsigned)crl_apb_rst_lpd_top | (23UL << 12),
+		[pctl_devreset_lpd_dbg_fpd] = (unsigned)crl_apb_rst_lpd_dbg | (0UL << 12),
+		[pctl_devreset_lpd_dbg_lpd] = (unsigned)crl_apb_rst_lpd_dbg | (1UL << 12),
+		[pctl_devreset_lpd_rpu_dbg0] = (unsigned)crl_apb_rst_lpd_dbg | (4UL << 12),
+		[pctl_devreset_lpd_rpu_dbg1] = (unsigned)crl_apb_rst_lpd_dbg | (5UL << 12),
+		[pctl_devreset_lpd_dbg_ack] = (unsigned)crl_apb_rst_lpd_dbg | (15UL << 12),
+		[pctl_devreset_fpd_sata] = (unsigned)crf_apb_rst_fpd_top | (1UL << 12),
+		[pctl_devreset_fpd_gt] = (unsigned)crf_apb_rst_fpd_top | (2UL << 12),
+		[pctl_devreset_fpd_gpu] = (unsigned)crf_apb_rst_fpd_top | (3UL << 12),
+		[pctl_devreset_fpd_gpu_pp0] = (unsigned)crf_apb_rst_fpd_top | (4UL << 12),
+		[pctl_devreset_fpd_gpu_pp1] = (unsigned)crf_apb_rst_fpd_top | (5UL << 12),
+		[pctl_devreset_fpd_fpd_dma] = (unsigned)crf_apb_rst_fpd_top | (6UL << 12),
+		[pctl_devreset_fpd_s_axi_hpc_0_fpd] = (unsigned)crf_apb_rst_fpd_top | (7UL << 12),
+		[pctl_devreset_fpd_s_axi_hpc_1_fpd] = (unsigned)crf_apb_rst_fpd_top | (8UL << 12),
+		[pctl_devreset_fpd_s_axi_hp_0_fpd] = (unsigned)crf_apb_rst_fpd_top | (9UL << 12),
+		[pctl_devreset_fpd_s_axi_hp_1_fpd] = (unsigned)crf_apb_rst_fpd_top | (10UL << 12),
+		[pctl_devreset_fpd_s_axi_hpc_2_fpd] = (unsigned)crf_apb_rst_fpd_top | (11UL << 12),
+		[pctl_devreset_fpd_s_axi_hpc_3_fpd] = (unsigned)crf_apb_rst_fpd_top | (12UL << 12),
+		[pctl_devreset_fpd_swdt] = (unsigned)crf_apb_rst_fpd_top | (15UL << 12),
+		[pctl_devreset_fpd_dp] = (unsigned)crf_apb_rst_fpd_top | (16UL << 12),
+		[pctl_devreset_fpd_pcie_ctrl] = (unsigned)crf_apb_rst_fpd_top | (17UL << 12),
+		[pctl_devreset_fpd_pcie_bridge] = (unsigned)crf_apb_rst_fpd_top | (18UL << 12),
+		[pctl_devreset_fpd_pcie_cfg] = (unsigned)crf_apb_rst_fpd_top | (19UL << 12),
+		[pctl_devreset_fpd_acpu0] = (unsigned)crf_apb_rst_fpd_apu | (0UL << 12),
+		[pctl_devreset_fpd_acpu1] = (unsigned)crf_apb_rst_fpd_apu | (1UL << 12),
+		[pctl_devreset_fpd_acpu2] = (unsigned)crf_apb_rst_fpd_apu | (2UL << 12),
+		[pctl_devreset_fpd_acpu3] = (unsigned)crf_apb_rst_fpd_apu | (3UL << 12),
+		[pctl_devreset_fpd_apu_l2] = (unsigned)crf_apb_rst_fpd_apu | (8UL << 12),
+		[pctl_devreset_fpd_acpu0_pwron] = (unsigned)crf_apb_rst_fpd_apu | (10UL << 12),
+		[pctl_devreset_fpd_acpu1_pwron] = (unsigned)crf_apb_rst_fpd_apu | (11UL << 12),
+		[pctl_devreset_fpd_acpu2_pwron] = (unsigned)crf_apb_rst_fpd_apu | (12UL << 12),
+		[pctl_devreset_fpd_acpu3_pwron] = (unsigned)crf_apb_rst_fpd_apu | (13UL << 12),
+		[pctl_devreset_fpd_ddr_apm] = (unsigned)crf_apb_rst_ddr_ss | (2UL << 12),
+		[pctl_devreset_fpd_ddr_reserved] = (unsigned)crf_apb_rst_ddr_ss | (3UL << 12),
 	};
 
 	if ((dev < pctl_devreset_lpd_gem0) || (dev > pctl_devreset_fpd_ddr_reserved)) {
